Extract KeyManager into its own file
